+++ scripts/install.py
@@ -1,6 +1,4 @@
 import launch
-from packaging import version
-import gradio
 
-if version.parse(gradio.__version__) < version.parse("3.30.0"):
-    launch.run_pip("install gradio==3.30.0", "requirements for Styles Editor")
+if not launch.is_installed("pyAesCrypt"):
+    launch.run_pip("install pyAesCrypt", "pyAesCrypt for Styles Editor")